refactor(web): split PersonalityLab, add error boundary, drop dead TrajectoryCanvas
