+++ apps/web/components/sections/PersonalityLab/presets.ts
@@ -1,0 +1,45 @@
+import { careful, distracted, fast, type PresetName, precise } from '@humanjs/core';
+
+export interface PersonalityMeta {
+  key: PresetName;
+  tagline: string;
+  description: string;
+  travelMs: number;
+  curvature: number;
+  jitterPercent: number;
+}
+
+export const personalityPresets: readonly PersonalityMeta[] = [
+  {
+    key: 'careful',
+    tagline: 'Reads everything twice.',
+    description: 'High dwell, medium curvature. Slow, deliberate trajectories.',
+    travelMs: careful.mouse.travelTimeMs,
+    curvature: careful.mouse.curvature,
+    jitterPercent: careful.mouse.travelTimeJitter,
+  },
+  {
+    key: 'fast',
+    tagline: 'Knows exactly where to go.',
+    description: 'Brisk travel, low dwell. Confident, no second-guessing.',
+    travelMs: fast.mouse.travelTimeMs,
+    curvature: fast.mouse.curvature,
+    jitterPercent: fast.mouse.travelTimeJitter,
+  },
+  {
+    key: 'distracted',
+    tagline: 'Multitasking. Lots of sidebars.',
+    description: 'Higher curvature and jitter, variable pauses between actions.',
+    travelMs: distracted.mouse.travelTimeMs,
+    curvature: distracted.mouse.curvature,
+    jitterPercent: distracted.mouse.travelTimeJitter,
+  },
+  {
+    key: 'precise',
+    tagline: 'Surgical.',
+    description: 'Near-straight paths, minimal noise. Direct intent.',
+    travelMs: precise.mouse.travelTimeMs,
+    curvature: precise.mouse.curvature,
+    jitterPercent: precise.mouse.travelTimeJitter,
+  },
+];
